[REFACTOR]profile.js
refactor block and mute
