+++ general.js
@@ -1,6 +1,16 @@
 const knex = require('./knex.js');
 const jwt = require("jsonwebtoken");
 
+const TYPE_BLOCK = 0;
+const TYPE_MUTE = 1;
+const TYPE_REPORT = 2;
+
+const typeMap = {
+    "block": TYPE_BLOCK,
+    "mute": TYPE_MUTE,
+    "report": TYPE_REPORT,
+};
+
 async function convert_our_id(user_id) {
     return knex('profile').pluck("id").where("user_id", user_id)
         .then(temp => {
@@ -59,8 +69,80 @@ async function define_id(test_id, res) {
     return id;
 }
 
+async function handleBlockAction(req, res, actionType) {
+    const ourid = await define_id(req.headers.authorization, res);
+    if (!ourid) return; // 인증 실패 시 종료
+
+    const { target_id } = req.body; // 이제 body에는 target_id만 있으면 됨
+
+    if (!target_id) {
+        return res.status(400).json({ success: false, message: "target_id가 필요합니다." });
+    }
+
+    if (!(actionType in typeMap)) {
+        return res.status(400).json({ success: false, message: "지원하지 않는 타입입니다." });
+    }
+
+    try {
+        await knex.transaction(async (trx) => {
+            // block_list에 등록
+            await trx("block_list").insert({
+                user_id: ourid, // JWT에서 뽑은 값
+                blocked_user_id: target_id,
+                type: typeMap[actionType],
+            });
+
+            // block일 경우 follow 관계를 backup + 삭제
+            if (actionType === "block") {
+                const isUserFollowTarget = await trx("follow")
+                    .where({ user_id: ourid, friend_id: target_id })
+                    .first();
+                const isTargetFollowUser = await trx("follow")
+                    .where({ user_id: target_id, friend_id: ourid })
+                    .first();
+
+                let relation = null;
+                if (isUserFollowTarget && isTargetFollowUser) relation = 2;
+                else if (isUserFollowTarget) relation = 0;
+                else if (isTargetFollowUser) relation = 1;
+
+                if (relation !== null) {
+                    // backup 테이블에 기록
+                    await trx("follow_backup").insert({
+                        user_id1: ourid,
+                        user_id2: target_id,
+                        relation,
+                    });
+
+                    // follow 테이블에서 삭제
+                    await trx("follow")
+                        .whereIn(["user_id", "friend_id"], [
+                            [ourid, target_id],
+                            [target_id, ourid],
+                        ])
+                        .del();
+                }
+            }
+        });
+
+        return res.json({ success: true, message: `${actionType} 등록 완료` });
+    } catch (err) {
+        if (err.errno === 1062) {
+            return res.status(409).json({ success: false, message: `이미 ${actionType}된 사용자입니다.` });
+        } else {
+            console.error(err);
+            return res.status(500).json({ success: false, message: "서버 오류" });
+        }
+    }
+}
+
 module.exports = {
     convert_our_id,
     define_id,
-    tmp_convert_our_id
+    tmp_convert_our_id,
+    handleBlockAction,
+    typeMap,
+    TYPE_BLOCK,
+    TYPE_MUTE,
+    TYPE_REPORT
 };
